devop: replace AppNetworkFee ui with SwapFee
